Fix Examples/{UIExplorer,Movies}
UIExplorerActivity.java and MoviesActivity.java should override
`getReactNativeHost` method.
