+++ ReactAndroid/src/main/java/com/facebook/react/ReactActivity.java
@@ -86,7 +86,7 @@ public abstract class ReactActivity extends Activity
    * determining this (default just checks {@code BuildConfig}).
    */
   protected boolean getUseDeveloperSupport() {
-    return ((ReactApplication) getApplication()).getReactNativeHost().getUseDeveloperSupport();
+    return getReactNativeHost().getUseDeveloperSupport();
   }
 
   @Override
